https://sonarcloud.io/project/issues?branch=dev&id=blish-hud_Blish-HUD&issues=AXsQXBRXATb_4uWK_Zu9&open=AXsQXBRXATb_4uWK_Zu9

+++ Blish HUD/GameServices/Input/Keyboard/TypedInputUtil.cs
@@ -49,7 +49,7 @@ namespace Blish_HUD.Input {
                 return "";
             }
 
-            int result = ToUnicode(vkCode, scanCode, keyState, output, (int)5, (uint)0);
+            int result = ToUnicode(vkCode, scanCode, keyState, output, 5, 0);
 
             switch (result) {
                 case -1:
